requirements.txt
beautifulsoup4==4.12.3
boto3==1.35.15
botocore==1.35.15
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.8
iniconfig==2.0.0
jmespath==1.0.1
lxml==5.3.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
soupsieve==2.6
urllib3==2.2.2
